requirements.txt
aiogram~=3.8.0
python-dotenv~=1.0.1
psycopg2~=2.9.9
APScheduler~=3.10.4
emoji~=2.12.1